postgres_to_es/etl/requirements.txt
python-dotenv
psycopg2-binary
pydantic==1.*
elasticsearch